requirements.txt
py-cord
python-dotenv
tortoise-orm
PyNaCl